+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.47.0](https://github.com/constructorfleet/forge/compare/forge-v0.46.0...forge-v0.47.0) (2026-09-05)
+
+
+### Features
+
+* TUI: no key scrolls the transcript by a page ([#660](https://github.com/constructorfleet/forge/issues/660)) ([f13fb46](https://github.com/constructorfleet/forge/commit/f13fb46a6cbb8d0e5217f11d0cb135c4d5de3c39)), closes [#573](https://github.com/constructorfleet/forge/issues/573)
+* TUI: order gate rows into the timeline by finish time ([#661](https://github.com/constructorfleet/forge/issues/661)) ([69a772c](https://github.com/constructorfleet/forge/commit/69a772c4dfce9d625bb9fab474455d81884a14f0)), closes [#530](https://github.com/constructorfleet/forge/issues/530)
+
 ## [0.46.0](https://github.com/constructorfleet/forge/compare/forge-v0.45.0...forge-v0.46.0) (2026-09-05)